Refactor codebase & add user auth
