+++ src/styles.scss
@@ -3,3 +3,38 @@
   padding: 0;
   box-sizing: border-box;
 }
+
+body {
+  display: flex;
+  flex-flow: row wrap;
+}
+
+nav{
+  width: 100%;
+  height: 15%;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.tree-list{
+  width: 30%;
+  padding: .5rem;
+}
+
+.branch-wrap{
+  width: 70%;
+  padding: .5rem;
+}
+
+.modal {
+  display: none;
+  position: fixed;
+  width: 40%;
+  height: auto;
+  top: 50%;
+  left: 50%;
+}
